Make read/unread mail-waiting stats easier to access
New JS User.stats properties:
- read_mail_waiting
- unread_mail_waiting
- spam_waiting

New @-codes:
- MAILR (read mail waiting)
- MAILU (unread mail waiting)

And the corresponding MAILR# and MAILU# codes (for non-current user's stats).

Addresses feature request #191.

+++ src/sbbs3/getmail.h
@@ -28,7 +28,7 @@
 extern "C" {
 #endif
 
-DLLEXPORT int		getmail(scfg_t* cfg, int usernumber, BOOL sent, uint16_t attr);
+DLLEXPORT int		getmail(scfg_t* cfg, int usernumber, BOOL sent, int attr);
 DLLEXPORT mail_t *	loadmail(smb_t* smb, uint32_t* msgs, uint usernumber
 							,int which, long mode);
 DLLEXPORT void		freemail(mail_t* mail);
